+++ python/runtime/pai/submitter.py
@@ -19,9 +19,9 @@ import string
 from os import path
 
 from runtime import db, oss
+from runtime.diagnostics import SQLFlowDiagnostic
 from runtime.pai import cluster_conf
 from runtime.pai.entry import tensorflow as tensorflow_entry
-from runtime.tensorflow.diag import SQLFlowDiagnostic
 
 LIFECYCLE_ON_TMP_TABLE = 7
 JOB_ARCHIVE_FILE = "job.tar.gz"
